refactor(model-server): make token envelopes lossless
Remove float32 log-probability encoding and expose one base64 mode that always
uses float64. This keeps token metadata bit-exact and avoids asking users to
choose a precision that can silently alter training data.

+++ responses_api_agents/simple_agent_with_compaction/tests/test_compaction.py
@@ -729,7 +729,7 @@ def test_capture_observed_completion_accepts_token_envelopes():
                 "content": "answer",
                 "prompt_token_ids": encode_token_list([1, 2], "i32"),
                 "generation_token_ids": encode_token_list([3, 4], "i32"),
-                "generation_log_probs": encode_token_list([-0.5, -0.25], "f32"),
+                "generation_log_probs": encode_token_list([-0.5, -0.25], "f64"),
             }
         ],
         rollout_id=history.rollout_id,
